requirements.txt
flask
opencv-python
numpy
pyexiv2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib